Update README.md to reinforce the need to read the release notes
New adopters of the Community Firmware continue to find and download the bundled firmware without reading the release notes carefully.  MAC users have also now developed issues unique to the MACOS, and they are not finding/reading the FAQs Wiki to diagnose their issues.

It would be helpful to make one more maintenance release, in which the readme instructions in the zip files were replaced by links to the wiki and to the release notes, and in which the display firmware files are also replaced by a URL to the Refactored v1.1 release files.