Update README.md to add pointers to the Wiki page
Since the information in the bundled zip release file readmes is now out of date, we have repeatedly needed to update both the online Readme and Wiki files in an attempt to capture and communicate the key troubleshooting and configuration info.  It seems only prudent and fair to update the Readme to explicitly acknowledge those facts and to point to those resources.  Prior to this, we tried to intercept our new users at the Release notes page, but users persist in not complying with RTFM guidelines.