Update README.md
After comparing the ./ini file between the CR6Community Firmware and the official release of Marlin 2.0.9.1, it was clear that the CR6Community Firmware differs significantly in its implementation of the STM32 build parameters.  There is no telling whether those are the only differences.  It only seems fair to add this caution to the head of the repository readme as "fair warning" to others....